old-tests/submission/W3C/domhtml/HTMLBodyElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-body-alink" rel="help">
<link rel="author" title="NIST">
</head>
<body alink="#0000ff"> 
<p>Hello, World</p>


<script>
var domtest = async_test("The aLink IDL attribute of the body element must reflect the element's alink content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valink;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("body");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      valink = testNode.aLink;

      assert_equals(valink,"#0000ff");
       
    domtest.done();
}));

</script>
<p>The aLink IDL attribute of the body element must reflect the element's alink content attribute.</p>
<div id="log"></div>
</body></html>
